+++ src/Reach/reachDigitalOut.h
@@ -13,7 +13,7 @@ class ReachDigitalOut: public ReachComponent{
         int get_subcomponent_number()override;
         int list_subcomponents(char *subComponentList)override;
         int get_instruction(char *instructionBuf)override;
-        int process_comand(uint8_t subComponent,uint8_t command,uint8_t value)override;
+        int process_comand(int subComponent,int command,int value)override;
     private:
         DigitalOut *digitalOutCollection[32];
         const int reachComponenetId = 1;
